+++ source/common/ipc/api.ts
@@ -61,6 +61,9 @@ import type {
   TrezorSignTransactionRequest,
   TrezorSignTransactionResponse,
   HardwareWalletConnectionRequest,
+  LedgerDevicePayload,
+  TrezorDevicePayload,
+  TrezorDeviceErrorPayload,
 } from '../types/hardware-wallets.types';
 
 /**
@@ -463,6 +466,7 @@ export type GetBlockSyncProgressMainResponse = {
 /**
  * Channels for connecting / interacting with Hardware Wallet devices
  */
+
 export const GET_HARDWARE_WALLET_TRANSPORT_CHANNEL =
   'GET_HARDWARE_WALLET_TRANSPORT_CHANNEL';
 export type getHardwareWalletTransportRendererRequest = HardwareWalletTransportDeviceRequest;
@@ -474,12 +478,16 @@ export type getExtendedPublicKeyMainResponse = HardwareWalletExtendedPublicKeyRe
 export const GET_CARDANO_ADA_APP_CHANNEL = 'GET_CARDANO_ADA_APP_CHANNEL';
 export type getCardanoAdaAppRendererRequest = {
   path: string | null | undefined;
+  product?: string;
 };
 export type getCardanoAdaAppMainResponse = HardwareWalletCardanoAdaAppResponse;
 export const GET_HARDWARE_WALLET_CONNECTION_CHANNEL =
   'GET_HARDWARE_WALLET_CONNECTION_CHANNEL';
 export type getHardwareWalletConnectionMainRequest = HardwareWalletConnectionRequest;
-export type getHardwareWalletConnectionRendererResponse = Record<string, any>;
+export type getHardwareWalletConnectionRendererResponse =
+  | LedgerDevicePayload
+  | TrezorDevicePayload
+  | TrezorDeviceErrorPayload;
 export const SIGN_TRANSACTION_LEDGER_CHANNEL =
   'SIGN_TRANSACTION_LEDGER_CHANNEL';
 export type signTransactionLedgerRendererRequest = LedgerSignTransactionRequest;
@@ -511,3 +519,7 @@ export type showAddressMainResponse = void;
 export const TOGGLE_RTS_FLAGS_MODE_CHANNEL = 'TOGGLE_RTS_FLAGS_MODE_CHANNEL';
 export type ToggleRTSFlagsModeRendererRequest = void;
 export type ToggleRTSFlagsModeMainResponse = void;
+export const DEVICE_NOT_CONNECTED = 'DEVICE_NOT_CONNECTED';
+export const WAIT_FOR_LEDGER_DEVICES = 'WAIT_FOR_LEDGER_DEVICES';
+export type waitForLedgerDevicesRequest = void;
+export type waitForLedgerDevicesResponse = LedgerDevicePayload;